test(runtime): isolate active bare throw rethrow regression
